backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["sudonya <daniilroitenberg@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
asyncpg = "^0.30.0"
python-dotenv = "^1.1.0"
sqlalchemy = "^2.0.40"
uvicorn = "^0.34.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
